redis/mcp_aws_server/requirements.txt
fastmcp
mcp
pyyaml
gunicorn
python-dotenv
boto3
markdownify